requirements.txt
# Python dependencies for Browsing Assistant
fastapi
uvicorn
httpx
python-dotenv
mcp>=1.9.4
